Get rid of allmydata.web.common_py3.
